// src/app/model/ObjectType.ts
export enum ObjectType {
  STRING = 'string',
  STRING_ARRAY = 'string_array',
  NUMBER = 'number',
  BOOLEAN = 'boolean',
  MEMBER = 'member',
  MEMBER_ARRAY = "member_array",
  NOTIFICATION_MEMBER = 'notificationMember',
  NOTIFICATION_MEMBER_ARRAY = 'notificationMember_array',
  PARTICIPANT = 'participant',
  PARTICIPANT_ARRAY = 'participant_array',
  ACTION = 'action',
  ACTION_ARRAY = 'action_array',
  MESSAGE_TYPE = 'messageType',
  MODULE = 'module',
  DOCUMENT = 'document',
  DOCUMENT_ARRAY = 'document_array',
  ROOM_STATE = 'roomState',
  ATTACHMENT = 'attachment',
  COMMENT = 'comment',
  COMMENT_ARRAY = 'comment_array',
  ACTOR_TYPE = 'actorType',
  ACTOR_STATE = 'actorState',
  IMAGE = 'image',
  FACET = 'facet',
  INVESTMENT = 'investment',
  FINANCIAL = 'financial',
  ADDRESS = 'address',
  FCM_TOKEN = 'fcmToken',
  FCM_TOKEN_ARRAY = 'fcmToken_array',
  TRL_STAGE = 'trlStage',
  PROJECTION_METHOD = 'projectionMethod',
  ECOSYSTEM_TYPE = 'ecosystemType',
  ECOSYSTEM_SERVICE_TYPE = 'ecosystemServiceType',
  ECOSYSTEM_SERVICE_TYPE_ARRAY = 'ecosystemServiceType_array',
  PERMISSION = 'permission',
  PERMISSION_ARRAY = 'permission_array',
  DEEP_TECH_CELL = 'deepTechCell',
  DEEP_TECH_CELL_ARRAY = 'deepTechCell_array',
  NOTIFICATION_CHANNEL = 'notificationChannel',
  NOTIFICATION_CHANNEL_ARRAY = 'notificationChannel_array',
  NOTIFICATION_TYPE = 'notificationType',
  EMAIL = 'email',

}